+++ Scripts/commentExtender.js
@@ -56,12 +56,26 @@ class CommentExtender {
             try {
                 text = editor.getTextInRange(docBlock);
             } catch(e) {
-                this.setContext({"return": true, "tab": true});
+                this.setContext({"return": false, "tab": false});
             }
 
             // Nova uses \uFFFC (Object Replacement Character) for
-            // snippet placeholders, if so set tab event to false
-            this.setContext({"return": true, "tab": (/[\uFFFC]+/.test(text) ? false : true)});
+            // snippet placeholders, if there is one within the
+            // docblock set tab event to false
+
+            let tabState = (/[\uFFFC]+/.test(text) ? false : true);
+
+            // Also ignore tabs for Ruby/Rust/Swift comments
+
+            if (
+                editor.document.syntax === "ruby" ||
+                editor.document.syntax === "rust" ||
+                editor.document.syntax === "swift"
+            ) {
+                tabState = false;
+            }
+
+            this.setContext({"return": true, "tab": tabState});
         }
     }
 
@@ -90,14 +104,19 @@ class CommentExtender {
 
             let indent = "";
 
-            if (/^\s*\/\*[*!]]/.test(lineText)) {
-                // cursor on docblock start line (/**)
+            if (/^\s*\/\*[*!]]/.test(lineText)) {                   // cursor on docblock start line (/** or /*!)
                 indent = lineText.replace(/^(\s*).*$/m, "$1 * ");
-            } else if (/^\s*\*(?!\/)/.test(lineText)) {
-                // cursor on line starting with asterisk
+
+            } else if (/^\s*\*(?!\/)/.test(lineText)) {             // cursor on line starting with asterisk
                 indent = lineText.replace(/^(\s*)\*(\s*).*$/m, "$1*$2");
-            } else {
-                // any other line
+
+            } else if (/^\s*\/{3}/.test(lineText)) {                // cursor on line starting with '///'
+                indent = lineText.replace(/^(\s*)\/{3}(\s*).*$/m, "$1///$2");
+
+            } else if (/^\s*#/.test(lineText)) {                    // cursor on line starting with '#'
+                indent = lineText.replace(/^(\s*)#(\s*).*$/m, "$1#$2");
+
+            } else {                                                // any other line
                 indent = lineText.replace(/^(\s*).*$/m, "$1");
             }
 
@@ -166,13 +185,38 @@ class CommentExtender {
     }
 
     /**
-     * Get ranges of all docBlocks in document (including incomplete blocks)
+     * Get ranges of all docBlocks in document
      */
     getDocBlockRanges(editor) {
-        const regex = new RegExp(
-            /^[\t ]*\/\*[*!](?:.)+?\*\/[\t ]*$/,
-            "gms"
-        );
+
+        let regex;
+        const syntax = editor.document.syntax;
+
+        switch(syntax) {
+        case "c":
+        case "cpp":
+        case "lsl":
+        case "objc":
+            regex = new RegExp(/^(?:[\t ]*\/\*[*!].+?\*\/[\t ]*)$/, "gms");
+            break;
+        case "java":
+        case "javascript":
+        case "jsx":
+        case "php":
+        case "typescript":
+        case "tsx":
+            regex = new RegExp(/^(?:[\t ]*\/\*\*.+?\*\/[\t ]*)$/, "gms");
+            break;
+        case "ruby":
+            regex = new RegExp(/^(?:[\t ]*#[^\n\r]*?[\n\r])/, "gms");
+            break;
+        case "rust":
+        case "swift":
+            regex = new RegExp(/^(?:[\t ]*\/{3}[^\n\r]*?[\n\r])/, "gms");
+            break;
+        default:
+            return null;
+        }
 
         const range = new Range(0, editor.document.length);
         const text = editor.getTextInRange(range);