+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## Version 0.9.0
+
+* **Note:** DocBlockr is now using language specific trigger chars, i.e.
+  * `/**` for Java/Javascript/PHP/etc.
+  * `/**` or `/*!` for C languages (depending on selected comment style)
+  * `///` for Rust/Swift
+  * `##` for Ruby
+* Added support for Swift *(experimental)*
+* Added support for Ruby *(experimental)*
+* Some Rust comment enhancements (following rust-lang RFC 1574)
+* Extension settings: Fixed a bug which prevented free-form text to be added to header comment blocks. Also, made the tab index for placeholders optional: `${placeholder}` and `${0:placeholder}` now both work.
+* More bugfixes
+
+
 ## Version 0.8.0
 
 * Added support for Java, ObjC/ObjC++ and Rust. *There may be issues!*
@@ -7,9 +21,9 @@
 
 ## Version 0.7.0
 
+* C/C++ support
 * Portuguese translation by [Gwyneth Llewelyn](https://github.com/GwynethLlewelyn)
 * German language additions ([Gwyneth Llewelyn](https://github.com/GwynethLlewelyn))
-* C/C++ support ... also requested by Gwyneth ;^)
 
 
 ## Version 0.6.1